\name{hiplarShow}
\alias{hiplarShow}
\title{Shows the crossover points for all functions}
\description{
	Shows crossover points for all the functions.
}

\usage{
hiplarShow()
}

\details{
	No specific details required. This function simply shows the user what the crossover points
	are for the different functions.
}
\examples{
	hiplarShow()
}
